requirements.txt
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
PyYAML>=6.0.1
requests>=2.32.0
SQLAlchemy>=2.0.30
psycopg[binary]>=3.2.0
streamlit>=1.35.0
pytest>=8.2.0
